AmalthAI_WebApp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
brotli
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
durationpy==0.9
flask==3.1.0
frozenlist==1.5.0
google-auth==2.39.0
grpcio==1.62.2
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
kubeflow-katib==0.17.0
markupsafe==3.0.2
multidict==6.3.2
oauthlib==3.2.2
pip==25.0.1
propcache==0.3.1
protobuf==4.25.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyjwt==2.10.1
pyopenssl==25.0.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
kubernetes
pyu2f==0.1.5
pyyaml==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==78.1.0
six==1.17.0
typing-extensions==4.13.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
werkzeug==3.1.3
wheel==0.45.1
yarl==1.18.3
zstandard==0.23.0

# pip-only section from YAML
annotated-types==0.7.0
beautifulsoup4==4.14.0
coloredlogs==15.0.1
contourpy==1.3.3
cvat-sdk==2.34.0
cycler==0.12.1
docker==7.1.0
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.0
future==1.0.0
gdown==5.2.0
humanfriendly==10.0
imageio==2.37.0
imgviz==1.7.6
kiwisolver==1.4.9
lazy-loader==0.4
loguru==0.7.3
matplotlib==3.10.6
mpmath==1.3.0
natsort==8.4.0
networkx==3.5
numpy==2.3.3
onnxruntime==1.23.0
osam==0.2.5
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
pydantic==2.11.9
pydantic-core==2.33.2
pyparsing==3.2.5
pyqt5==5.15.11
pyqt5-qt5==5.15.17
pyqt5-sip==12.17.0
scikit-image==0.25.2
scipy==1.16.2
soupsieve==2.8
sympy==1.14.0
tifffile==2025.9.20
tinydb==4.8.2
tqdm==4.67.1
tuspy==0.2.5
typing-inspection==0.4.1